requirements.txt
# WEBAG Professional - Requirements
# Core Flask dependencies
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1

# Security and validation
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
bleach==6.1.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter

# GIS and file processing
# togeojson==0.1.1  # Package not available, will implement manually if needed

# Production server
gunicorn==21.2.0

# Development and testing (optional)
pytest==7.4.3
pytest-flask==1.3.0